+++ src/wasm/wasm-emscripten.cpp
@@ -311,15 +311,6 @@ Function* EmscriptenGlueGenerator::generateMemoryGrowthFunction() {
   return growFunction;
 }
 
-void EmscriptenGlueGenerator::generateStackInitialization(Address addr) {
-  auto* stackPointer = getStackPointerGlobal();
-  assert(!stackPointer->imported());
-  if (!stackPointer->init || !stackPointer->init->is<Const>()) {
-    Fatal() << "stack pointer global is not assignable";
-  }
-  stackPointer->init->cast<Const>()->value = Literal(int32_t(addr));
-}
-
 inline void exportFunction(Module& wasm, Name name, bool must_export) {
   if (!wasm.getFunctionOrNull(name)) {
     assert(!must_export);
@@ -405,6 +396,32 @@ private:
   std::unique_ptr<Builder> builder;
   Global* stackPointer;
 };
+
+// lld can sometimes produce a build with an imported mutable __stack_pointer
+// (i.e.  when linking with -fpie).  This method internalizes the
+// __stack_pointer and initializes it from an immutable global instead.
+// For -shared builds we instead call replaceStackPointerGlobal.
+void EmscriptenGlueGenerator::internalizeStackPointerGlobal() {
+  Global* stackPointer = getStackPointerGlobal();
+  if (!stackPointer || !stackPointer->imported() || !stackPointer->mutable_) {
+    return;
+  }
+
+  Name internalName = stackPointer->name;
+  Name externalName = internalName.c_str() + std::string("_import");
+
+  // Rename the imported global, and make it immutable
+  stackPointer->name = externalName;
+  stackPointer->mutable_ = false;
+  wasm.updateMaps();
+
+  // Create a new global with the old name that is not imported.
+  Builder builder(wasm);
+  auto* init = builder.makeGlobalGet(externalName, stackPointer->type);
+  auto* sp = builder.makeGlobal(
+    internalName, stackPointer->type, init, Builder::Mutable);
+  wasm.addGlobal(sp);
+}
 
 void EmscriptenGlueGenerator::replaceStackPointerGlobal() {
   Global* stackPointer = getStackPointerGlobal();